+++ bin/docfd.ml
@@ -406,6 +406,7 @@ let run
     ~(env : Eio_unix.Stdenv.base)
     ~sw
     (debug_log : string option)
+    (scan_hidden : bool)
     (max_depth : int)
     (max_fuzzy_edit_dist : int)
     (max_token_search_dist : int)
@@ -470,6 +471,7 @@ let run
             )
         )
     );
+  Params.scan_hidden := scan_hidden;
   Params.max_file_tree_scan_depth := max_depth;
   Params.max_fuzzy_edit_dist := max_fuzzy_edit_dist;
   Params.max_token_search_dist := max_token_search_dist;
@@ -1150,6 +1152,7 @@ let cmd ~env ~sw =
   Cmd.v (Cmd.info "docfd" ~version ~doc)
     (const (run ~env ~sw)
      $ debug_log_arg
+     $ hidden_arg
      $ max_depth_arg
      $ max_fuzzy_edit_dist_arg
      $ max_token_search_dist_arg